chore: Refactor proto info reader (#115)
